requirements.txt
# Install in virtual environment with:
# pip install -r requirements.txt

flask==2.3.3
python-dotenv==1.0.0
langchain
langchain-openai
langchain-core
langchain-community
pydantic
flask-cors

# Add any other dependencies here
